repositories_added

+++ src/lab3/classes/Teacher.java
@@ -4,20 +4,20 @@ import java.util.List;
 
 public class Teacher extends Person{
 
-    private int teacherId;
+    private long teacherId;
     private List<Course> courses;
 
-    public Teacher(String firstName, String lastName, int teacherId, List<Course> courses) {
+    public Teacher(String firstName, String lastName, long teacherId, List<Course> courses) {
         super(firstName, lastName);
         this.teacherId = teacherId;
         this.courses = courses;
     }
 
-    public int getTeacherId() {
+    public long getTeacherId() {
         return teacherId;
     }
 
-    public void setTeacherId(int teacherId) {
+    public void setTeacherId(long teacherId) {
         this.teacherId = teacherId;
     }
